requirements.txt
Pillow>=11.0.0
nltk==3.9.1
torch==2.8.0
pandas==2.3.1
tensorflow==2.20.0rc0
fastapi==0.116.1
peft==0.17.0
transformers==4.55.0
tqdm==4.67.1
bitsandbytes
python-multipart
uvicorn